♻️ 공용 API dto 사용하는 방식으로 변경
