+++ app/backend/src/groups/dto/create-groups.dto.ts
@@ -1,0 +1,5 @@
+import { RequestGroupsDto } from '@morak/apitype/dto/request/groups';
+
+export class CreateGroupsDto implements RequestGroupsDto {
+  title: string;
+}
