+++ app/backend/src/member/dto/member.dto.ts
@@ -1,5 +1,5 @@
+import { ResponseMemberDto } from '@morak/apitype/dto/response/member';
 import { ApiProperty } from '@nestjs/swagger';
-import { ResponseMemberDto } from '@morak/apitype/DTO/response/member'
 
 export class MemberInformationDto implements ResponseMemberDto {
   @ApiProperty({ description: 'Provider ID of the user', example: '123456' })
@@ -13,18 +13,4 @@ export class MemberInformationDto implements ResponseMemberDto {
 
   @ApiProperty({ description: "URL of the user's profile picture", example: 'https://example.com/profile.jpg' })
   profilePicture: string;
-}
-
-export class MemberDto {
-  @ApiProperty({ description: 'ID of the Mogaco', example: '3' })
-  id: bigint;
-
-  @ApiProperty({ description: 'ProviderId of the Mogaco', example: '123456' })
-  providerId: string;
-
-  @ApiProperty({ description: 'email of the Mogaco', example: 'bcwm.morak@gmail.com' })
-  email: string;
-
-  @ApiProperty({ description: 'socialType of the Mogaco', example: 'google' })
-  socialType: string;
 }
