--- app/backend/src/mogaco/dto/index.ts
@@ -1,2 +1,0 @@
-export * from './create-mogaco.dto';
-export * from './mogaco.dto';
